foodgram_backend/requirements.txt
Django==3.2.13
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary
python-dotenv
djoser
pillow
drf-extra-fields
django-debug-toolbar
reportlab
django-filter==2.4.0
